+++ src/lib/pages/interaction/component/ModuleSelectDrawerButton.tsx
@@ -1,6 +1,5 @@
 import {
   Button,
-  Text,
   Heading,
   useDisclosure,
   Drawer,
@@ -9,13 +8,16 @@ import {
   DrawerHeader,
   DrawerCloseButton,
   DrawerBody,
-  Flex,
 } from "@chakra-ui/react";
 
-import { ExplorerLink } from "lib/components/ExplorerLink";
 import { CustomIcon } from "lib/components/icon";
-import { LabelText } from "lib/components/LabelText";
-import { StateImage } from "lib/components/state";
+
+import { ModuleSelectBody } from "./select-module/ModuleSelectBody";
+import {
+  ModuleSelectorDisplay,
+  ModuleSelectorInput,
+} from "./select-module/ModuleSeletor";
+// import { ModuleEmptyState } from "./select-module/ModuleEmptyState";
 
 interface ModuleSelectDrawerButtonProps {
   buttonText?: string;
@@ -52,75 +54,12 @@ export const ModuleSelectDrawerButton = ({
             display="flex"
             flexDirection="column"
           >
-            <Flex bgColor="gray.800" px={4} py={3} mb={6} borderRadius={4}>
-              {/* Input */}
-              {/* <Flex justifyContent="space-between" w="full">
-                TODO Input here
-                <Flex gap={2}>
-                  <Button
-                    variant="primary"
-                    size="sm"
-                    onClick={() => {
-                      onOpen();
-                    }}
-                  >
-                    Submit
-                  </Button>
-                  <Button
-                    variant="outline-white"
-                    size="sm"
-                    onClick={() => {
-                      onOpen();
-                    }}
-                  >
-                    Cancel
-                  </Button>
-                </Flex>
-              </Flex> */}
-              {/* Selected Address */}
-              <Flex justifyContent="space-between" w="full" alignItems="center">
-                <LabelText flex="1" label="Viewing Address">
-                  <ExplorerLink
-                    value="cltn1m9y7um59yxtmek68qkwg3ykm28s52rrell6prx"
-                    type="user_address"
-                    textFormat="normal"
-                  />
-                </LabelText>
-                <LabelText flex="1" label="Hex">
-                  <ExplorerLink
-                    value="0xe688b84b23f322a994A53dbF8E15FA82CDB71127"
-                    type="user_address"
-                    textFormat="normal"
-                  />
-                </LabelText>
-                <Button
-                  variant="outline-white"
-                  size="sm"
-                  onClick={() => {
-                    onOpen();
-                  }}
-                >
-                  <CustomIcon name="swap" boxSize={3} />
-                  Change Address
-                </Button>
-              </Flex>
-            </Flex>
-            <Flex
-              h="full"
-              justifyContent="center"
-              alignItems="center"
-              border="1px solid"
-              borderRadius={4}
-              borderColor="gray.700"
-            >
-              {/* Empty State */}
-              <Flex flexDirection="column" gap={4} alignItems="center">
-                <StateImage imageVariant="empty" width="160px" />
-                <Text variant="body2" color="text.dark">
-                  Available functions for selected modules will display here
-                </Text>
-              </Flex>
-            </Flex>
+            {/* Input */}
+            <ModuleSelectorInput />
+            {/* Selected Address */}
+            <ModuleSelectorDisplay />
+            <ModuleSelectBody />
+            {/* <ModuleEmptyState /> */}
           </DrawerBody>
         </DrawerContent>
       </Drawer>